Revert "Install common bundles via config model [run-systemtest]"
