+++ config-model/src/test/java/com/yahoo/vespa/model/admin/metricsproxy/MetricsProxyContainerClusterTest.java
@@ -51,7 +51,8 @@ public class MetricsProxyContainerClusterTest {
         var builder = new PlatformBundlesConfig.Builder();
         model.getConfig(builder, CLUSTER_CONFIG_ID);
         PlatformBundlesConfig config = builder.build();
-        assertThat(config.bundlePaths(), hasItem(endsWith(METRICS_PROXY_BUNDLE_FILE.toString())));
+        assertEquals(1, config.bundlePaths().size());
+        assertThat(config.bundlePaths(0), endsWith(METRICS_PROXY_BUNDLE_FILE.toString()));
     }
 
     @Test
