+++ config-model/src/test/java/com/yahoo/vespa/model/container/xml/BundleInstantiationSpecificationBuilderTest.java
@@ -5,7 +5,6 @@ import com.yahoo.component.ComponentSpecification;
 import com.yahoo.config.model.builder.xml.XmlHelper;
 import com.yahoo.container.bundle.BundleInstantiationSpecification;
 import com.yahoo.search.grouping.GroupingValidator;
-import com.yahoo.vespa.model.container.PlatformBundles;
 import org.junit.Test;
 import org.w3c.dom.Element;
 import org.xml.sax.SAXException;
