+++ container-disc/pom.xml
@@ -176,14 +176,12 @@
                 jdisc-security-filters-jar-with-dependencies.jar,
                 vespaclient-container-plugin-jar-with-dependencies.jar,
                 vespa-athenz-jar-with-dependencies.jar,
-                container-apache-http-client-bundle-jar-with-dependencies.jar, <!-- Apache http client repackaged as bundle -->
-
-                <!-- Vespa security utils with necessary 3rd party bundles -->
                 security-utils-jar-with-dependencies.jar,
-                bcpkix-jdk15on-${bouncycastle.version}.jar,
-                bcprov-jdk15on-${bouncycastle.version}.jar,
-
-               <!-- Jetty -->
+                zkfacade-jar-with-dependencies.jar,
+                zookeeper-server-jar-with-dependencies.jar,
+                <!-- Apache http client repackaged as bundle -->
+                container-apache-http-client-bundle-jar-with-dependencies.jar,
+                <!-- Jetty -->
                 alpn-api-${jetty-alpn.version}.jar,
                 http2-server-${jetty.version}.jar,
                 http2-common-${jetty.version}.jar,
@@ -200,8 +198,6 @@
                 jetty-servlets-${jetty.version}.jar,
                 jetty-util-${jetty.version}.jar,
                 jetty-util-ajax-${jetty.version}.jar,
-                javax.servlet-api-3.1.0.jar,
-
                 <!-- Spifly (required for OSGi service loader used by Jetty)  -->
                 org.apache.aries.spifly.dynamic.bundle-${spifly.version}.jar,
                 asm-${asm.version}.jar,
@@ -210,7 +206,10 @@
                 asm-tree-${asm.version}.jar,
                 asm-util-${asm.version}.jar,
                 <!-- Spifly end -->
-
+                <!-- Misc 3rd party bundles -->
+                bcpkix-jdk15on-${bouncycastle.version}.jar,
+                bcprov-jdk15on-${bouncycastle.version}.jar,
+                javax.servlet-api-3.1.0.jar,
                 <!-- Jersey 2 + Jackson 2 -->
                 aopalliance-repackaged-${hk2.version}.jar,
                 hk2-api-${hk2.version}.jar,
